Bug fixes and route changes
I have removed some comments from index.js as they were using up
the number of lines. Also moved the login check to /upload, as we want
users to be able to be able to do everything but upload without logging
in. Updated /upload to work with updated extractTags method. Added
check to see if current user is the owner of a song in the detail view.