Tweaked upload form
I have modified the song upload form to only accept mp3 files for the
song and only image files for the album art. I have also added an
attribute to the song file input to make it required, to make it less
difficult to check whether a user uploaded a file or not. Also made the
playlist selection filed necessary as a song needs to be uploaded to
a playlist.